bot/requirements.txt
python-telegram-bot==21.5
google-genai
anthropic==0.34.0
httpx==0.27.2
tweepy==4.14.0
requests==2.32.3
aiohttp==3.10.5